requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
langchain==0.3.24
langchain-community==0.3.22
langchain-core==0.3.55
mysql-connector-python==9.3.0
requests==2.32.3
streamlit==1.41.1
requests==2.32.3
